+++ src/main/java/de/samply/frontend/dto/configuration/ProjectConfigurationMatcher.java
@@ -1,5 +1,6 @@
 package de.samply.frontend.dto.configuration;
 
+import de.samply.app.ProjectManagerConst;
 import de.samply.frontend.dto.Project;
 
 import java.lang.reflect.Field;
@@ -11,18 +12,25 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 public class ProjectConfigurationMatcher {
 
+    private static Map<String, Project> CUSTOM_CONFIGURATION = Map.of(ProjectManagerConst.CUSTOM_PROJECT_CONFIGURATION, new Project());
+
     public static Map<String, Project> fetchMatchProjectConfiguration(Project project, Map<String, Project> projectConfigurations) throws ProjectConfigurationMatcherException {
+        if (project.isCustomConfig()) {
+            return CUSTOM_CONFIGURATION;
+        }
         Map<Map.Entry<String, Project>, Integer> matchedFieldCounts = new HashMap<>();
         for (Map.Entry<String, Project> entry : projectConfigurations.entrySet()) {
-            int matchedFields = countMatchedFields(project, entry.getValue());
-            if (matchedFields > 0) {
-                matchedFieldCounts.put(entry, matchedFields);
+            if (!entry.getKey().equals(ProjectManagerConst.CUSTOM_PROJECT_CONFIGURATION)) {
+                int matchedFields = countMatchedFields(project, entry.getValue());
+                if (matchedFields > 0) {
+                    matchedFieldCounts.put(entry, matchedFields);
+                }
             }
         }
         Optional<Map.Entry<String, Project>> nameProjectConfig = matchedFieldCounts.entrySet().stream()
                 .max(Comparator.comparingInt(Map.Entry::getValue))
                 .map(Map.Entry::getKey);
-        return (nameProjectConfig.isPresent()) ? Map.ofEntries(nameProjectConfig.get()) : new HashMap<>();
+        return (nameProjectConfig.isPresent()) ? Map.ofEntries(nameProjectConfig.get()) : CUSTOM_CONFIGURATION;
     }
 
     private static int countMatchedFields(Project project, Project projectConfiguration) {